Test short timings for buzzer
