+++ src/events.py
@@ -9,6 +9,7 @@ import updateserver
 import os
 
 from lock import config_lock
+from src import GPIOconfig
 
 path = os.path.dirname(os.path.abspath(__file__))
 
@@ -828,7 +829,7 @@ def button_detects_change(gpio, level, tick):
         logger.info(f"{E1} push button1 is pressed at " + str(datetime.now()))
         mag_E1_allowed_to_open = True
         relay.trigger_relay_one(E1_thirdPartyOption)
-        relay.test_buzzer(8, 1, 1)
+        GPIOconfig.test_buzzer(8, 1, 1)
         eventsMod.record_button_pressed(E1, "Security Guard Button")
 
     elif gpio == E2_Button:
